+++ packages/flutter_query/lib/flutter_query.dart
@@ -57,6 +57,8 @@ export 'src/core/core.dart'
         RefetchOnResume,
         RefetchOnReconnect,
         RetryResolver,
+        Seed,
+        SeedUpdatedAt,
         ShouldRebuild,
 
         // Cancellation